stt_gen_summary/requirements.txt
streamlit
google-cloud-aiplatform
Pillow
opencv-python
google-cloud-speech
pyyaml
